backend/requirements.txt
fastapi
uvicorn[standard]
snowflake-connector-python
pandas
langchain
openai
faiss-cpu
python-dotenv
